Move NoteUtils.setTags() method to CardTemplatePreviewer
Only used in CardTemplatePreviewer, the now empty Noteutils file was
deleted.
